% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{analyse_chord}
\alias{analyse_chord}
\title{Analyse a Chord Specified as Scale Degrees}
\usage{
analyse_chord(left = character(), right = character())
}
\arguments{
\item{left, right}{character vectors of scale degrees for the left and
right hand.}
}
\value{
A tibble with one row and the following columns:
\itemize{
\item \code{rooted}: is it a rooted chord, i.e., is the first note in the left hand
the root.
\item \code{left_hand}: is it a left-handed chord, i.e., is the right hand empty.
\item \code{top_degree}: scale degree of the top note in the chord. For a left handed chord,
this is the top note of the left hand.
\item \code{alterations}: character vector of alterations used in the chord, sorted in
increasing order.
}
}
\description{
Analyse a chord that is specified as scale degrees for the left and
right hand.
}
