% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{flatten}
\alias{flatten}
\alias{sharpen}
\title{Flatten or Sharpen a Note}
\usage{
flatten(notes)

sharpen(notes)
}
\arguments{
\item{notes}{a character vector of valid note names.}
}
\value{
a character vector of valid note names that are either a half-tone
lower or higher than \code{notes}.
}
\description{
A character vector of note names is flattened or sharpened. Double
accidentals are converted to notes in the C major scales.
}
