% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{get_enharmonic_equivalent}
\alias{get_enharmonic_equivalent}
\title{Get the Enharmonic Equivalent of a Note}
\usage{
get_enharmonic_equivalent(notes)
}
\arguments{
\item{notes}{a vector of valid note names.}
}
\value{
character vector with note names representing the enharmonic
equivalents of \code{notes}.
}
\description{
This function returns the enharmonic equivalent of notes. Notes with
accidentals are either changed to a note with the other type of
accidental (e.g., F# => Gb, Bb => A#) or to a note without
accidental (e.g., Cb => B, E# => F). Notes without accidentals are
left unchanged unless there is a enharmonic equivalent with one
accidental (e.g., B => Cb, F => E#).
}
