% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano-package.R
\docType{data}
\name{gsa_results}
\alias{gsa_results}
\title{Gene set analysis result data}
\format{A list where each element is an object of class GSAres, as returned
by \code{runGSA}.}
\description{
This data set contains gene set analysis results, as returned by the
\code{\link{runGSA}} function, that is used as example data for downstream
functions. The input data to \code{runGSA} was randomly generated and is
accessible through \code{data(gsa_input)}.
}
\keyword{datasets}
