% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{is_valid_chord}
\alias{is_valid_chord}
\title{Check Plausibility of the Scale Degrees in a Chord}
\usage{
is_valid_chord(
  left = character(),
  right = character(),
  type = c("major", "minor", "dominant", "half-diminished", "diminished", "augmented")
)
}
\arguments{
\item{left, right}{character vectors of scale degrees for the left and
right hand.}

\item{type}{the type of chord that was entered. One of "major", "minor",
"dominant", "half-diminished", "diminished", or "augmented". The type determines
which degrees are invalid. See 'Details' for more information.}
}
\value{
a logical indicating whether the chord only contains valid scale degrees.
}
\description{
Check that the scale degrees contained in a chord are plausible for the
given type. For example, a chord of type "dominant" should not contain
b3 or 7. The function only checks the absence of implausible scale degrees,
but not the presence of expected ones. So, a "dominant" chord is considered
valid even if 3 and/or b7 are missing. This allows to, say, also define a
sus4 chord as a "dominant" chord. See 'Details' for more information.
}
\details{
In general, the following scale degrees are allowed in chords:
1, 2, b3, 3, 4, b5, 5, #5, 6, b7, 7, b9, 9, #9, 11, #11, b13, 13.

Depending on the chord type, some degrees are not allowed:
\itemize{
\item "major": b3, b7, b5
\item "minor": 3
\item "dominant": b3, 7
\item "half-diminished": 3, 5, 7
\item "diminished": 3, 5, 7
\item "augmented": b3, 5
}
}
