% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{notes_toupper}
\alias{notes_toupper}
\alias{notes_tolower}
\title{Convert Note Names to Uppercase or Lowercase}
\usage{
notes_toupper(notes)

notes_tolower(notes)
}
\arguments{
\item{notes}{character vector of notes. The function does not check
whether these are valid names. The functions can also process keys, but this will
change their meaning, because, e.g., "C1" and "c1" are not the same key.}
}
\value{
a character vector with transformed note names

@export
}
\description{
Convert note names to upper case or lower case and ensure that the accidental
"b" remains lower case.
}
