% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_piano}
\alias{plot_piano}
\title{Plot a Piano Keyboard}
\usage{
plot_piano(
  lower = "A2",
  upper = "c5",
  labels = c("none", "white", "black", "all"),
  black_labels = c("sharp", "flat")
)
}
\arguments{
\item{lower, upper}{the lowest and highest key of the piano keyboard
to plot. Both values must correspond to a white key, i.e. a note of the
C major scale, between A2 and c5 and \code{upper} must be at least one key above
\code{lower}.}

\item{labels}{character indicating whether labels with note names should
be plotted. Use \code{"white"} or \code{"black"} to only show the labels for white
or black keys, respectively and \code{"all"} to show labels for all keys.}

\item{black_labels}{character indicating the names to be used for the labels
of the black keys.}
}
\value{
a \code{ggplot} object
}
\description{
Plot a Piano Keyboard
}
