% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{print.chords_db}
\alias{print.chords_db}
\title{Print Chords Database}
\usage{
\method{print}{chords_db}(x, width = NULL, n = NULL, ...)
}
\arguments{
\item{x}{chords database to print, usually created with \code{\link[=read_chords_csv]{read_chords_csv()}}.}

\item{width}{Width of text output to generate. This defaults to \code{NULL}, which
means use the \code{width} \link[pillar:pillar_options]{option}.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than the \code{print_max} \link[pillar:pillar_options]{option}.
Otherwise, will print as many rows as specified by the
\code{print_min} \link[pillar:pillar_options]{option}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Print a chords database created by \code{\link[=read_chords_csv]{read_chords_csv()}}. This basically converts
list columns to a readable format and then prints the tibble in the usual way.
The function thus also respects all the options described in
\link[pillar:pillar_options]{pillar::pillar_options}.
}
