% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_chords_csv}
\alias{read_chords_csv}
\title{Read a Chord Database from a csv File}
\usage{
read_chords_csv(file)
}
\arguments{
\item{file}{the path to a csv file containing the chord database. The column
delimiter must be a comma (\verb{,}). See 'Details' for more information on the
file format.}
}
\value{
a tibble with one row for each chord. It contains all the columns from the csv file
("type", "notation", "name", "left", "right") and the column that are returned
by \code{\link[=analyse_chord]{analyse_chord()}} ("rooted", "left_hand", "top_degree", "alterations").
}
\description{
Read a Chord Database from a csv File
}
\details{
The csv file must contain the column titles in the first row. Each row contains
the data for one chord voicing. Note that voicings are defined by their scale
degrees not by specific notes or even keys. The file must contain all the required
columns, but some values are allowed to be empty. The order of the columns is
irrelevant and additional columns are allowed but will be ignored in the output.

The following lists the 5 required columns and specifies whether a value must
be provided or is optional:
\itemize{
\item "type" (required): the type of chord, one of "major", "minor", "dominant",
"half-diminished", "diminished" or "augmented". The function checks that a valid
value is passed and verifies that the scale degrees match the type
(see \code{\link[=is_valid_chord]{is_valid_chord()}}).
\item "notation" (optional): notation for the chord (e.g., "maj7" or "7"). This is simply
used as an identifier and any value is allowed.
\item "name" (optional): a name for the voicing. This is simply used as an identifier
and any value is allowed.
\item "left", "right" (required): the scale degrees for the left and right hand
separated by spaces (e.g., "3 b7 9"). At least on of the two must have a value,
but it is allowed to leave the other empty. This allows to define left-hand
voicings. The function checks that all scale degrees are valid and plausible
for the given chord type.
}
}
