% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees.R
\name{sort_degrees}
\alias{sort_degrees}
\title{Sort a Vector of Scale Degrees}
\usage{
sort_degrees(degrees, decreasing = FALSE)
}
\arguments{
\item{degrees}{character vector of scale degrees. The function does not
check whether all scale degrees are valid for performance reasons.}

\item{decreasing}{logical indicating the sort order. Default is \code{FALSE}
which implies increasing order.}
}
\value{
the sorted character of scale degrees.
}
\description{
Sort a character vector of scale degrees.
}
