% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{degrees_to_notes}
\alias{degrees_to_notes}
\title{Convert Scale Degrees to Notes}
\usage{
degrees_to_notes(degrees, root_note = "C", scale = NULL)
}
\arguments{
\item{degrees}{character indicating the degrees degrees to include in
the left and right hand, respectively. Valid values are the degrees degrees
returned by \code{\link[=get_major_scale_with_alt]{get_major_scale_with_alt()}}: all integers from 1 to 15 and
the alterations b3, b5, #5, b7, b9, #9, #11, and b13.}

\item{root_note}{a single note indicating the root note of the chord.}

\item{scale}{the output of \code{\link[=get_major_scale_with_alt]{get_major_scale_with_alt()}} for the key corresponding
to the desired root note. This is mainly provided to improve performance, if many
chords with the same root need to be computed, since \code{get_major_scale_with_alt()}
is where the functions spends most its time.}
}
\description{
Convert a vector of scale degrees to notes for a given root note.
}
