% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{flatten}
\alias{flatten}
\alias{sharpen}
\title{Flatten or Sharpen a Note}
\usage{
flatten(notes, use_double_accidentals = TRUE)

sharpen(notes, use_double_accidentals = TRUE)
}
\arguments{
\item{notes}{a character vector of valid note names.}

\item{use_double_accidentals}{should double accidentals be used, i.e.,
should a second accidental be added if there already is one.}
}
\value{
a character vector of valid note names that are either a half-tone
lower or higher than \code{notes}.
}
\description{
A character vector of note names is flattened or sharpened. Double
accidentals are converted to notes in the C major scales.
}
