% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{get_chord}
\alias{get_chord}
\title{Generate a Chord from Intervals}
\usage{
get_chord(left = c(), right = c(), root = "C")
}
\arguments{
\item{left, right}{character indicating the intervals to include in
the left and right hand, respectively. Valid values are the interval names
returned by \code{\link[=get_major_scale_with_alt]{get_major_scale_with_alt()}}: all integers from 1 to 15 and
the alterations b3, b5, #5, b7, b9, #9, #11, and b13.}

\item{root}{Either a single note name indicating the root key, or the output
of \code{\link[=get_major_scale_with_alt]{get_major_scale_with_alt()}} for a precomputed scale.}
}
\value{
A list with components \code{left} and \code{right}, each containing a
character vector of note names.
}
\description{
Create a chord voicing by selecting intervals from a major scale with
alterations.
}
