% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{get_enharmonic_equivalent}
\alias{get_enharmonic_equivalent}
\title{Get the Enharmonic Equivalent of a Note}
\usage{
get_enharmonic_equivalent(
  notes,
  use_double_accidentals = c("none", "sharp", "flat")
)
}
\arguments{
\item{notes}{a vector of valid note names.}

\item{use_double_accidentals}{logical indicating whether enharmonic
equivalents involving double accidentals (i.e. D => C##) should be
returned. For white keys that have an enharmonic equivalent with
only accidental that one is preferred independent of the setting of
\code{use_double_accidentals}. E.g., for C, B# is preferred over Dbb.}
}
\value{
character vector with note names representing the enharmonic
equivalents of \code{notes}.
}
\description{
This function returns the enharmonic equivalent of notes. Notes with
accidentals are either changed to a note with the other type of
accidental (e.g., F# => Gb, Bb => A#) or to a note without
accidental (e.g., Cb => B, E# => F). Notes without accidentals are
left unchanged unless there is a enharmonic equivalent with one
accidental (e.g., B => Cb, F => E#).
}
