% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_major_scale}
\alias{get_major_scale}
\title{Get the Notes of a Major Scale}
\usage{
get_major_scale(key, n_octave = 1L)
}
\arguments{
\item{key}{character indicating the key. Supported are
keys with up to six accidentals.}

\item{n_octave}{integer indicating the number of octaves to
return.}
}
\description{
Get the Notes of a Major Scale
}
