\name{gsa_results}
     \docType{data}
     \alias{gsa_results}
     \title{Gene set analysis result data}
     \description{
       This data set contains gene set analysis results, as returned by the \code{\link{runGSA}} function, that is used as example data for downstream functions. The input data to \code{runGSA} was randomly generated and is accessible through \code{data(gsa_input)}.
     }
     \usage{data(gsa_results)}
     \format{A list where each element is an object of class GSAres, as returned by \code{runGSA}.}

     \keyword{datasets}
