% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{has_accidental}
\alias{has_accidental}
\title{Does a Note Name Contain an Accidental}
\usage{
has_accidental(notes, which = c("any", "sharp", "flat"))
}
\arguments{
\item{notes}{a vector of notes names}

\item{which}{character indicating the type of accidental to look for.
One of "any", "sharp", "flat".}
}
\value{
a logical vector indicating whether the note names in \code{notes}
contain an accidental.
}
\description{
Does a Note Name Contain an Accidental
}
