% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{notes_toupper}
\alias{notes_toupper}
\alias{notes_tolower}
\title{Convert Note Names to Uppercase or Lowercase}
\usage{
notes_toupper(notes)

notes_tolower(notes)
}
\arguments{
\item{notes}{character vector of note names. The function does not check
whether these are valid names.}
}
\value{
a character vector with transformed note names

@export
}
\description{
Convert note names to upper case or lower case and ensure that the accidental
"b" remains lower case.
}
