% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano-package.R
\docType{package}
\name{piano-package}
\alias{piano}
\alias{piano-package}
\title{Visualise a Piano Keyboard}
\description{
Visualise a piano keyboard with ggplot2.
}
\author{
\strong{Maintainer}: Stefan Lanz \email{slanz1137@gmail.com}

}
\keyword{internal}
